feat(dashboard): add react-grid-layout implementation
- Add react-grid-layout dependency for professional grid management
- Add dashboard-container-rgl.tsx with full RGL implementation
- Add widget-wrapper-rgl.tsx for RGL-compatible widget wrapper
- Preserve legacy implementation in dashboard-container-legacy.tsx
- Convert dashboard-container.tsx to router between implementations
- Add custom Tailwind styles for react-grid-layout elements

The RGL implementation provides:
- Responsive breakpoints (lg: 1200px, md: 996px, sm: 768px)
- Discrete size snapping per widget type
- Proper bin-packing for varying widget heights
- Drag handle with visual feedback
