+++ package.json
@@ -83,7 +83,9 @@
     "react": "^18.3.1",
     "react-day-picker": "^9.11.0",
     "react-dom": "^18.3.1",
+    "react-grid-layout": "^2.2.2",
     "react-hook-form": "^7.63.0",
+    "react-resizable": "^3.1.3",
     "react-window": "^1.8.8",
     "recharts": "^2.15.4",
     "tailwind-merge": "^2.3.0",
@@ -98,6 +100,7 @@
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
     "@types/node": "^24.5.2",
+    "@types/react-grid-layout": "^1.3.6",
     "@types/uuid": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^7.16.0",
     "@typescript-eslint/parser": "^7.16.0",
